% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTE.R
\name{der.LtLik}
\alias{der.LtLik}
\title{Derivatives of Tangent Likelihood w.r.t. beta}
\usage{
der.LtLik(X, u, e, p, t0)
}
\arguments{
\item{X}{design matrix}

\item{u}{density values}

\item{e}{eesidual}

\item{p}{Taylor expansion order}

\item{t0}{tangent point}
}
\value{
It returns to first and second derivatives of tangent-likelihood w.r.t. beta.
This function is called in MTE main function.
}
\description{
Derivatives of Tangent Likelihood w.r.t. beta
}
\keyword{internal}
