% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Huber.R
\name{huberloss}
\alias{huberloss}
\title{Huber Loss}
\usage{
huberloss(r, alpha)
}
\arguments{
\item{r}{residual, y-Xbeta}

\item{alpha}{1/alpha is the huber tuning parameter delta. Larger alpha results in smaller portion of squared loss.}
}
\value{
it returns huber loss that will be called in Huber estimation.
}
\description{
Huber Loss
}
