% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTE.R
\name{lnt}
\alias{lnt}
\title{Tangent-likelihood function.}
\usage{
lnt(f, t, p)
}
\arguments{
\item{f}{values of density function.}

\item{t}{tangent point. It is positive and close to 0. If 0, the tangent-likelihood is equivalent to log-likelihood.}

\item{p}{Taylor expansion order, the value can be set up to 3.}
}
\description{
The function calculates tangent-likelihood given the value of density.
}
\examples{
set.seed(2017)
x=c(rnorm(80), rnorm(20, 10, 10))
f=dnorm(x)
y=lnt(f, 0.1, 2)
}
